refactor: 운동 상세 조회 훅을 entities/exercise로 통합
- pages/session, pages/exercise-detail이 GET /catalog/exercises/{id}를
  각자 쿼리키·타입으로 따로 구현하고 있던 걸 entities/exercise 하나로 합친다
- exercise-detail/api/types.ts가 스웨거 생성 타입(ExerciseDetailResponse/
  MuscleResponse/VoiceCueResponse)을 손으로 다시 선언해둔 것도 정리 —
  스웨거에 아직 없는 stepOrder/totalStepCount만 로컬로 확장해서 남긴다
