+++ src/entities/exercise/index.ts
@@ -1,0 +1,2 @@
+export { exerciseQueryKey } from "./model/exercise";
+export { useExercise } from "./model/use-exercise";
