+++ src/entities/exercise/model/exercise.ts
@@ -1,0 +1,3 @@
+export function exerciseQueryKey(exerciseId: number | null) {
+  return ["catalog", "exercises", exerciseId] as const;
+}
